refactor : userController에서 HTTP 관련 처리하게 수정 및 주석 수정
