+++ src/main/java/corecord/dev/domain/user/application/UserService.java
@@ -36,7 +36,6 @@ import java.util.regex.Pattern;
 public class UserService {
 
     private final JwtUtil jwtUtil;
-    private final CookieUtil cookieUtil;
     private final RefreshTokenRepository refreshTokenRepository;
     private final AnalysisDbService analysisDbService;
     private final AbilityDbService abilityDbService;
@@ -45,23 +44,19 @@ public class UserService {
     private final FolderDbService folderDbService;
     private final RecordDbService recordDbService;
 
-
-    @Value("${jwt.access-token.expiration-time}")
-    private long accessTokenExpirationTime;
-
-    @Value("${jwt.refresh-token.expiration-time}")
-    private long refreshTokenExpirationTime;
-
+    /**
+     * 회원가입
+     * @param registerToken
+     * @param userRegisterDto
+     * @return
+     */
     @Transactional
-    public UserResponse.UserDto registerUser(HttpServletResponse response, String registerToken, UserRequest.UserRegisterDto userRegisterDto) {
-        // registerToken 유효성 검증
+    public UserResponse.UserDto registerUser(String registerToken, UserRequest.UserRegisterDto userRegisterDto) {
         validRegisterToken(registerToken);
-
-        // user 정보 유효성 검증
         validateUserInfo(userRegisterDto.getNickName());
 
         String providerId = jwtUtil.getProviderIdFromToken(registerToken);
-        // 이미 존재하는 유저인지 확인
+
         checkExistUser(providerId);
 
         // 새로운 유저 생성
@@ -72,23 +67,27 @@ public class UserService {
         String refreshToken = jwtUtil.generateRefreshToken(savedUser.getUserId());
         saveRefreshToken(refreshToken, savedUser);
 
-        // AccessToken 및 RefreshToken 쿠키 설정
-        setTokenCookies(response, "accessToken", jwtUtil.generateAccessToken(savedUser.getUserId()));
-        setTokenCookies(response, "refreshToken", refreshToken);
-        return UserConverter.toUserDto(savedUser);
+        String accessToken = jwtUtil.generateAccessToken(savedUser.getUserId());
+
+        return UserConverter.toUserDto(savedUser, accessToken, refreshToken);
     }
 
-    // 로그아웃
+    /**
+     * 로그아웃
+     * @param refreshToken
+     */
     @Transactional
-    public void logoutUser(HttpServletRequest request, HttpServletResponse response) {
-        // RefreshToken 삭제
-        deleteRefreshTokenInRedis(request);
-        deleteTokenCookies(response);
+    public void logoutUser(String refreshToken) {
+        deleteRefreshTokenInRedis(refreshToken);
     }
 
-    // 유저 삭제
+    /**
+     * 회원 탈퇴
+     * @param userId
+     * @param refreshToken
+     */
     @Transactional
-    public void deleteUser(HttpServletRequest request, HttpServletResponse response, Long userId) {
+    public void deleteUser(Long userId, String refreshToken) {
         // 연관된 데이터 삭제
         abilityDbService.deleteAbilityByUserId(userId);
         analysisDbService.deleteAnalysisByUserId(userId);
@@ -97,13 +96,14 @@ public class UserService {
         chatDbService.deleteChatRoomByUserId(userId);
         folderDbService.deleteFolderByUserId(userId);
         userDbService.deleteUserByUserId(userId);
-
-        // RefreshToken 삭제
-        deleteRefreshTokenInRedis(request);
-        deleteTokenCookies(response);
+        deleteRefreshTokenInRedis(refreshToken);
     }
 
-    // 유저 정보 수정
+    /**
+     * 유저 정보 수정
+     * @param userId
+     * @param userUpdateDto
+     */
     @Transactional
     public void updateUser(Long userId, UserRequest.UserUpdateDto userUpdateDto) {
         User user = userDbService.getUser(userId);
@@ -118,7 +118,11 @@ public class UserService {
         }
     }
 
-    // 유저 정보 조회
+    /**
+     * 유저 정보 조회
+     * @param userId
+     * @return
+     */
     @Transactional
     public UserResponse.UserInfoDto getUserInfo(Long userId) {
         User user = userDbService.getUser(userId);
@@ -127,37 +131,17 @@ public class UserService {
         return UserConverter.toUserInfoDto(user, recordCount);
     }
 
-    // RefreshToken 저장
     private void saveRefreshToken(String refreshToken, User user) {
         RefreshToken newRefreshToken = RefreshToken.of(refreshToken, user.getUserId());
         refreshTokenRepository.save(newRefreshToken);
     }
 
-    // 토큰 쿠키 설정
-    private void setTokenCookies(HttpServletResponse response, String tokenName, String token) {
-        if(tokenName.equals("accessToken")) {
-            ResponseCookie accessTokenCookie = cookieUtil.createTokenCookie(tokenName, token, accessTokenExpirationTime);
-            response.addHeader("Set-Cookie", accessTokenCookie.toString());
-        } else {
-            ResponseCookie refreshTokenCookie = cookieUtil.createTokenCookie(tokenName, token, refreshTokenExpirationTime);
-            response.addHeader("Set-Cookie", refreshTokenCookie.toString());
-        }
-    }
-
-    private void deleteTokenCookies(HttpServletResponse response) {
-        ResponseCookie accessTokenCookie = cookieUtil.deleteCookie("accessToken");
-        response.addHeader("Set-Cookie", accessTokenCookie.toString());
-        ResponseCookie refreshTokenCookie = cookieUtil.deleteCookie("refreshToken");
-        response.addHeader("Set-Cookie", refreshTokenCookie.toString());
-    }
-
-    private void deleteRefreshTokenInRedis(HttpServletRequest request) {
-        String refreshToken = cookieUtil.getCookieValue(request, "refreshToken");
+    private void deleteRefreshTokenInRedis(String refreshToken) {
         if (refreshToken == null || refreshToken.isEmpty()) {
             log.info("쿠키에 리프레쉬 토큰 없음");
             return;
         }
-        Optional<RefreshToken> refreshTokenOptional = refreshTokenRepository.findByRefreshToken(cookieUtil.getCookieValue(request, "refreshToken"));
+        Optional<RefreshToken> refreshTokenOptional = refreshTokenRepository.findByRefreshToken(refreshToken);
         refreshTokenOptional.ifPresent(refreshTokenRepository::delete);
     }
 
@@ -167,7 +151,6 @@ public class UserService {
         }
     }
 
-    // user 정보 유효성 검증
     private void validateUserInfo(String nickName) {
         if (nickName == null || nickName.isEmpty() || nickName.length() > 10) {
             throw new UserException(UserErrorStatus.INVALID_USER_NICKNAME);
@@ -180,7 +163,6 @@ public class UserService {
         }
     }
 
-    // registerToken 유효성 검증
     private void validRegisterToken(String registerToken) {
         if (!jwtUtil.isRegisterTokenValid(registerToken)) {
             throw new TokenException(TokenErrorStatus.INVALID_REGISTER_TOKEN);
