+++ src/main/java/corecord/dev/domain/user/presentation/UserController.java
@@ -2,6 +2,7 @@ package corecord.dev.domain.user.presentation;
 
 import corecord.dev.common.response.ApiResponse;
 import corecord.dev.common.status.SuccessStatus;
+import corecord.dev.common.util.CookieUtil;
 import corecord.dev.common.web.UserId;
 import corecord.dev.domain.user.status.UserSuccessStatus;
 import corecord.dev.domain.user.domain.dto.request.UserRequest;
@@ -10,6 +11,8 @@ import corecord.dev.domain.user.application.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,13 +21,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/users")
 public class UserController {
     private final UserService userService;
+    private final CookieUtil cookieUtil;
 
-    @GetMapping("/test")
-    public ResponseEntity<ApiResponse<String>> getSuccess(
-            @UserId Long userId
-    ) {
-        return ApiResponse.success(SuccessStatus.OK, "userId: " + userId);
-    }
+    @Value("${jwt.access-token.expiration-time}")
+    private long accessTokenExpirationTime;
+
+    @Value("${jwt.refresh-token.expiration-time}")
+    private long refreshTokenExpirationTime;
 
     @PostMapping("/register")
     public ResponseEntity<ApiResponse<UserResponse.UserDto>> registerUser(
@@ -32,7 +35,10 @@ public class UserController {
             @RequestHeader("registerToken") String registerToken,
             @RequestBody UserRequest.UserRegisterDto userRegisterDto
             ) {
-        UserResponse.UserDto registerResponse = userService.registerUser(response, registerToken, userRegisterDto);
+        UserResponse.UserDto registerResponse = userService.registerUser(registerToken, userRegisterDto);
+
+        createTokenCookies(response, registerResponse.getAccessToken(), registerResponse.getRefreshToken());
+
         return ApiResponse.success(UserSuccessStatus.USER_REGISTER_SUCCESS, registerResponse);
     }
 
@@ -41,7 +47,10 @@ public class UserController {
             HttpServletRequest request,
             HttpServletResponse response
     ) {
-        userService.logoutUser(request, response);
+        String refreshToken = cookieUtil.getCookieValue(request, "refreshToken");
+        userService.logoutUser(refreshToken);
+        deleteTokenCookies(response);
+
         return ApiResponse.success(UserSuccessStatus.USER_LOGOUT_SUCCESS);
     }
 
@@ -51,7 +60,10 @@ public class UserController {
             HttpServletResponse response,
             @UserId Long userId
     ) {
-        userService.deleteUser(request, response, userId);
+        String refreshToken = cookieUtil.getCookieValue(request, "refreshToken");
+        userService.deleteUser(userId, refreshToken);
+        deleteTokenCookies(response);
+
         return ApiResponse.success(UserSuccessStatus.USER_DELETE_SUCCESS);
     }
 
@@ -70,5 +82,19 @@ public class UserController {
     ) {
         UserResponse.UserInfoDto userInfoDto = userService.getUserInfo(userId);
         return ApiResponse.success(UserSuccessStatus.GET_USER_INFO_SUCCESS, userInfoDto);
+    }
+
+    private void createTokenCookies(HttpServletResponse response, String accessToken, String refreshToken) {
+        ResponseCookie accessTokenCookie = cookieUtil.createTokenCookie("accessToken", accessToken, accessTokenExpirationTime);
+        response.addHeader("Set-Cookie", accessTokenCookie.toString());
+        ResponseCookie refreshTokenCookie = cookieUtil.createTokenCookie("refreshToken", refreshToken, refreshTokenExpirationTime);
+        response.addHeader("Set-Cookie", refreshTokenCookie.toString());
+    }
+
+    private void deleteTokenCookies(HttpServletResponse response) {
+        ResponseCookie accessTokenCookie = cookieUtil.deleteCookie("accessToken");
+        response.addHeader("Set-Cookie", accessTokenCookie.toString());
+        ResponseCookie refreshTokenCookie = cookieUtil.deleteCookie("refreshToken");
+        response.addHeader("Set-Cookie", refreshTokenCookie.toString());
     }
 }
